add -user management- page

+++ src/main/java/com/example/Project/repository/IUserRepository.java
@@ -1,9 +1,11 @@
 package com.example.Project.repository;
 
-import com.example.Project.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface IUserRepository extends JpaRepository<User, Long> {
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.Project.model.User;
+
+public interface IUserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByUsername(String username);
 }
